requirements.txt
certifi==2020.12.5
click==7.1.2
entrypoints==0.3
jedi==0.17.0
joblib==1.0.1
MarkupSafe==1.1.1
mistune==0.8.4
nltk==3.5
numpy>=1.9.2
pandas>=0.19
pytz==2021.1
scipy>=0.15.1
scikit-learn>=0.18
terminado==0.9.2
threadpoolctl==2.1.0
tqdm==4.59.0
webencodings==0.5.1
wincertstore==0.2
Flask==1.1.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
Werkzeug==0.15.5
regex==2020.11.13
tensorflow_text==2.9.0
lime==0.2.0.1